Fix/bugs (#458)
* Bump version number

* Fix bug in live loss plot callback

* Minor fixes

* Fix bug with state key strings

* Update CHANGELOG.md
