+++ torchbearer/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2.5'
+__version__ = '0.2.6.dev'
